Fields in sub-folders like /modal must have the same casing as the type variable
