+++ libraries/j2commerce/src/Field/Modal/ContactMultiselectField.php
@@ -10,7 +10,7 @@
 
 namespace J2Commerce\Library\J2Commerce\Field\Modal;
 
-use J2Commerce\Library\J2Commerce\Field\ModalMultiSelectField;
+use J2Commerce\Library\J2Commerce\Field\ModalMultiselectField;
 use Joomla\CMS\Factory;
 use Joomla\CMS\Language\LanguageHelper;
 use Joomla\CMS\Language\Text;
@@ -22,14 +22,14 @@ use Joomla\CMS\Uri\Uri;
 // phpcs:enable PSR1.Files.SideEffects
 
 /**
- * Library version of UserMultiField for multi-user selection.
+ * Library version of ContactMultiSelectField for multi-contact selection.
  *
- * This field provides a modal interface for selecting multiple users from com_users.
+ * This field provides a modal interface for selecting multiple contacts from com_contact.
  * It can be used by any extension that includes the J2Commerce library.
  *
  * @since  1.0.0
  */
-class UserMultiSelectField extends ModalMultiSelectField
+class ContactMultiselectField extends ModalMultiselectField
 {
     /**
      * The form field type.
@@ -37,7 +37,7 @@ class UserMultiSelectField extends ModalMultiSelectField
      * @var     string
      * @since   1.0.0
      */
-    protected $type = 'Modal_UserMultiselect';
+    protected $type = 'Modal_ContactMultiselect';
 
     /**
      * Method to attach a Form object to the field.
@@ -53,7 +53,7 @@ class UserMultiSelectField extends ModalMultiSelectField
      */
     public function setup(\SimpleXMLElement $element, $value, $group = null)
     {
-        // Handle comma-separated values for multiple users
+        // Handle comma-separated values for multiple contacts
         if ($value && \is_string($value) && strpos($value, ',') !== false) {
             $values = explode(',', $value);
             $value  = array_map('intval', array_filter($values));
@@ -73,11 +73,11 @@ class UserMultiSelectField extends ModalMultiSelectField
         // Prepare enabled actions
         $this->canDo['propagate'] = ((string) $this->element['propagate'] == 'true') && \count($languages) > 2;
 
-        // Prepare Urls - use the multi-select modal layout from com_users
+        // Prepare Urls - use the multi-select modal layout from com_contact
         $linkItems = (new Uri())->setPath(Uri::base(true) . '/index.php');
         $linkItems->setQuery([
-            'option'                => 'com_users',
-            'view'                  => 'users',
+            'option'                => 'com_contact',
+            'view'                  => 'contacts',
             'layout'                => 'modal_multiselect',
             'tmpl'                  => 'component',
             'function'              => 'jSelectItemMultiCallback_' . $this->id,
@@ -86,17 +86,17 @@ class UserMultiSelectField extends ModalMultiSelectField
 
         if ($language) {
             $linkItems->setVar('forcedLanguage', $language);
-            $modalTitle                            = Text::_('LIB_J2COMMERCE_SELECT_USERS') . ' &#8212; ' . $this->getTitle();
+            $modalTitle                            = Text::_('LIB_J2COMMERCE_SELECT_CONTACTS') . ' &#8212; ' . $this->getTitle();
             $this->dataAttributes['data-language'] = $language;
         } else {
-            $modalTitle = Text::_('LIB_J2COMMERCE_USER_SELECT_MODAL_ADD_USERS');
+            $modalTitle = Text::_('LIB_J2COMMERCE_CONTACT_SELECT_MODAL_ADD_CONTACTS');
         }
 
         $this->urls['select']        = (string) $linkItems;
         $this->modalTitles['select'] = $modalTitle;
-        $this->hint                  = $this->hint ?: Text::_('LIB_J2COMMERCE_SELECT_USERS');
+        $this->hint                  = $this->hint ?: Text::_('LIB_J2COMMERCE_SELECT_CONTACTS');
 
-        $this->sql_title_table  = '#__users';
+        $this->sql_title_table  = '#__contact_details';
         $this->sql_title_column = 'name';
         $this->sql_title_key    = 'id';
 
@@ -124,27 +124,27 @@ class UserMultiSelectField extends ModalMultiSelectField
         // Get the parent input HTML
         $html = parent::getInput();
 
-        // Add a container for the users table
+        // Add a container for the contacts table
         $html .= '<div id="' . $this->id . '_table">';
 
-        // Build the table with the selected users on first load
+        // Build the table with the selected contacts on first load
         if (!empty($this->value)) {
-            // Fetch user titles from the database
-            $users = $this->getValueTitles();
-            $html .= '<div class="my-2"><strong>' . Text::_('LIB_J2COMMERCE_SELECTED_USERS') . ' (' . \count($this->value) . '):</strong></div>';
-            $html .= '<table class="table table-sm table-striped"><thead><tr><th class="w-10">' . Text::_('LIB_J2COMMERCE_USER_FIELD_ID') . '</th><th>' . Text::_('LIB_J2COMMERCE_USER_FIELD_NAME') . '</th><th class="text-end w-6"><button type="button" class="btn btn-sm btn-outline-danger" onclick="clearAllItems_' . $this->id . '()" title="' . Text::_('LIB_J2COMMERCE_USERS_CLEAR_ALL') . '"><span class="icon-trash" aria-hidden="true"></span></button></th><th class="w-1"><span class="visually-hidden">' . Text::_('LIB_J2COMMERCE_REMOVE') . '</span></th></tr></thead><tbody>';
-            foreach ($this->value as $index => $userId) {
-                $userName = isset($users[$userId]) ? htmlspecialchars($users[$userId]->name, ENT_QUOTES, 'UTF-8') : $userId;
+            // Fetch contact titles from the database
+            $contacts = $this->getValueTitles();
+            $html .= '<div class="my-2"><strong>' . Text::_('LIB_J2COMMERCE_SELECTED_CONTACTS') . ' (' . \count($this->value) . '):</strong></div>';
+            $html .= '<table class="table table-sm table-striped"><thead><tr><th class="w-10">' . Text::_('LIB_J2COMMERCE_CONTACT_FIELD_ID') . '</th><th>' . Text::_('LIB_J2COMMERCE_CONTACT_FIELD_NAME') . '</th><th class="text-end w-6"><button type="button" class="btn btn-sm btn-outline-danger" onclick="clearAllItems_' . $this->id . '()" title="' . Text::_('LIB_J2COMMERCE_CONTACTS_CLEAR_ALL') . '"><span class="icon-trash" aria-hidden="true"></span></button></th><th class="w-1"><span class="visually-hidden">' . Text::_('LIB_J2COMMERCE_REMOVE') . '</span></th></tr></thead><tbody>';
+            foreach ($this->value as $index => $contactId) {
+                $contactName = isset($contacts[$contactId]) ? htmlspecialchars($contacts[$contactId]->name, ENT_QUOTES, 'UTF-8') : $contactId;
                 $html .= '<tr>';
-                $html .= '<td class="fw-bold">' . $userId . '</td>';
-                $html .= '<td>' . $userName . '</td>';
-                $html .= '<td class="text-end"><button type="button" class="btn btn-sm btn-danger" onclick="removeItem_' . $this->id . '(' . $userId . ')" title="' . Text::_('LIB_J2COMMERCE_USER_CLEAR') . '"><span class="icon-trash" aria-hidden="true"></span></button></td>';
-                $html .= '<td class="w-1"><input type="hidden" name="jform[request][user_ids][' . $index . ']" value="' . $userId . '" id="' . $this->id . '_hidden_' . $index . '" data-title="' . $userName . '"></td>';
+                $html .= '<td class="fw-bold">' . $contactId . '</td>';
+                $html .= '<td>' . $contactName . '</td>';
+                $html .= '<td class="text-end"><button type="button" class="btn btn-sm btn-danger" onclick="removeItem_' . $this->id . '(' . $contactId . ')" title="' . Text::_('LIB_J2COMMERCE_CONTACT_CLEAR') . '"><span class="icon-trash" aria-hidden="true"></span></button></td>';
+                $html .= '<td class="w-1"><input type="hidden" name="jform[request][contact_ids][' . $index . ']" value="' . $contactId . '" id="' . $this->id . '_hidden_' . $index . '" data-title="' . $contactName . '"></td>';
                 $html .= '</tr>';
             }
             $html .= '</tbody></table>';
         } else {
-            $html .= '<div class="my-2">' . Text::_('LIB_J2COMMERCE_NO_USERS_SELECTED') . '</div>';
+            $html .= '<div class="my-2">' . Text::_('LIB_J2COMMERCE_NO_CONTACTS_SELECTED') . '</div>';
         }
         $html .= '</div>';
 
@@ -171,11 +171,11 @@ class UserMultiSelectField extends ModalMultiSelectField
         );
 
         // Load language strings for JavaScript
-        Text::script('LIB_J2COMMERCE_SELECTED_USERS');
-        Text::script('LIB_J2COMMERCE_USER_FIELD_ID');
-        Text::script('LIB_J2COMMERCE_USER_FIELD_NAME');
-        Text::script('LIB_J2COMMERCE_USERS_CLEAR_ALL');
-        Text::script('LIB_J2COMMERCE_USER_CLEAR');
+        Text::script('LIB_J2COMMERCE_SELECTED_CONTACTS');
+        Text::script('LIB_J2COMMERCE_CONTACT_FIELD_ID');
+        Text::script('LIB_J2COMMERCE_CONTACT_FIELD_NAME');
+        Text::script('LIB_J2COMMERCE_CONTACTS_CLEAR_ALL');
+        Text::script('LIB_J2COMMERCE_CONTACT_CLEAR');
         Text::script('LIB_J2COMMERCE_REMOVE');
 
         // Initialize this specific field instance via inline script
@@ -184,7 +184,7 @@ class UserMultiSelectField extends ModalMultiSelectField
             if (window.initItemMultiField) {
                 const handler = window.initItemMultiField('{$this->id}');
 
-                // Set up user-specific table update function
+                // Set up contact-specific table update function
                 handler.customUpdateTable = function(selectedItems) {
                     const tableContainer = document.getElementById('{$this->id}_table');
                     if (!tableContainer) {
@@ -199,7 +199,7 @@ class UserMultiSelectField extends ModalMultiSelectField
                         // Create table structure
                         const caption = handler.createElement('div', {
                             className: 'my-2',
-                            innerHTML: '<strong>' + Joomla.Text._('LIB_J2COMMERCE_SELECTED_USERS') + ' (' + selectedItems.length + '):</strong>'
+                            innerHTML: '<strong>' + Joomla.Text._('LIB_J2COMMERCE_SELECTED_CONTACTS') + ' (' + selectedItems.length + '):</strong>'
                         });
 
                         const table = handler.createElement('table', {
@@ -208,12 +208,12 @@ class UserMultiSelectField extends ModalMultiSelectField
 
                         const thead = handler.createElement('thead');
                         const headerRow = handler.createElement('tr', {
-                            innerHTML: '<th class=\"w-10\">' + Joomla.Text._('LIB_J2COMMERCE_USER_FIELD_ID') + '</th>' +
-                                      '<th>' + Joomla.Text._('LIB_J2COMMERCE_USER_FIELD_NAME') + '</th>' +
+                            innerHTML: '<th class=\"w-10\">' + Joomla.Text._('LIB_J2COMMERCE_CONTACT_FIELD_ID') + '</th>' +
+                                      '<th>' + Joomla.Text._('LIB_J2COMMERCE_CONTACT_FIELD_NAME') + '</th>' +
                                       '<th class=\"text-end w-6\">' +
                                           '<button type=\"button\" class=\"btn btn-sm btn-outline-danger\" ' +
                                                   'onclick=\"clearAllItems_{$this->id}()\" ' +
-                                                  'title=\"' + Joomla.Text._('LIB_J2COMMERCE_USERS_CLEAR_ALL') + '\">' +
+                                                  'title=\"' + Joomla.Text._('LIB_J2COMMERCE_CONTACTS_CLEAR_ALL') + '\">' +
                                               '<span class=\"icon-trash\" aria-hidden=\"true\"></span>' +
                                           '</button>' +
                                       '</th>' +
@@ -231,7 +231,7 @@ class UserMultiSelectField extends ModalMultiSelectField
                             // Create hidden field
                             const hiddenField = handler.createElement('input', {
                                 type: 'hidden',
-                                name: 'jform[request][user_ids][' + index + ']',
+                                name: 'jform[request][contact_ids][' + index + ']',
                                 value: item.id,
                                 id: '{$this->id}_hidden_' + index
                             });
@@ -245,7 +245,7 @@ class UserMultiSelectField extends ModalMultiSelectField
                                           '<td class=\"text-end\">' +
                                               '<button type=\"button\" class=\"btn btn-sm btn-danger\" ' +
                                                       'onclick=\"removeItem_{$this->id}(' + item.id + ')\" ' +
-                                                      'title=\"' + Joomla.Text._('LIB_J2COMMERCE_USER_CLEAR') + '\">' +
+                                                      'title=\"' + Joomla.Text._('LIB_J2COMMERCE_CONTACT_CLEAR') + '\">' +
                                                   '<span class=\"icon-trash\" aria-hidden=\"true\"></span>' +
                                               '</button>' +
                                           '</td>';
@@ -281,7 +281,7 @@ class UserMultiSelectField extends ModalMultiSelectField
         $data                          = parent::getLayoutData();
         $data['language']              = (string) $this->element['language'];
         $data['multiple']              = true;
-        $data['buttonIcons']['select'] = 'fa-solid fa-users';
+        $data['buttonIcons']['select'] = 'fa-solid fa-address-book';
 
         return $data;
     }
