+++ libraries/j2commerce/src/Field/Modal/UserMultiselectField.php
@@ -10,7 +10,7 @@
 
 namespace J2Commerce\Library\J2Commerce\Field\Modal;
 
-use J2Commerce\Library\J2Commerce\Field\ModalMultiSelectField;
+use J2Commerce\Library\J2Commerce\Field\ModalMultiselectField;
 use Joomla\CMS\Factory;
 use Joomla\CMS\Language\LanguageHelper;
 use Joomla\CMS\Language\Text;
@@ -22,14 +22,14 @@ use Joomla\CMS\Uri\Uri;
 // phpcs:enable PSR1.Files.SideEffects
 
 /**
- * Library version of ArticleMultiSelectField for multi-article selection.
+ * Library version of UserMultiField for multi-user selection.
  *
- * This field provides a modal interface for selecting multiple articles from com_content.
+ * This field provides a modal interface for selecting multiple users from com_users.
  * It can be used by any extension that includes the J2Commerce library.
  *
  * @since  1.0.0
  */
-class ArticleMultiSelectField extends ModalMultiSelectField
+class UserMultiselectField extends ModalMultiselectField
 {
     /**
      * The form field type.
@@ -37,7 +37,7 @@ class ArticleMultiSelectField extends ModalMultiSelectField
      * @var     string
      * @since   1.0.0
      */
-    protected $type = 'Modal_ArticleMultiselect';
+    protected $type = 'Modal_UserMultiselect';
 
     /**
      * Method to attach a Form object to the field.
@@ -53,7 +53,7 @@ class ArticleMultiSelectField extends ModalMultiSelectField
      */
     public function setup(\SimpleXMLElement $element, $value, $group = null)
     {
-        // Handle comma-separated values for multiple articles
+        // Handle comma-separated values for multiple users
         if ($value && \is_string($value) && strpos($value, ',') !== false) {
             $values = explode(',', $value);
             $value  = array_map('intval', array_filter($values));
@@ -73,11 +73,11 @@ class ArticleMultiSelectField extends ModalMultiSelectField
         // Prepare enabled actions
         $this->canDo['propagate'] = ((string) $this->element['propagate'] == 'true') && \count($languages) > 2;
 
-        // Prepare Urls - use the multi-select modal layout from com_content
+        // Prepare Urls - use the multi-select modal layout from com_users
         $linkItems = (new Uri())->setPath(Uri::base(true) . '/index.php');
         $linkItems->setQuery([
-            'option'                => 'com_content',
-            'view'                  => 'articles',
+            'option'                => 'com_users',
+            'view'                  => 'users',
             'layout'                => 'modal_multiselect',
             'tmpl'                  => 'component',
             'function'              => 'jSelectItemMultiCallback_' . $this->id,
@@ -86,18 +86,18 @@ class ArticleMultiSelectField extends ModalMultiSelectField
 
         if ($language) {
             $linkItems->setVar('forcedLanguage', $language);
-            $modalTitle                            = Text::_('LIB_J2COMMERCE_SELECT_ARTICLES') . ' &#8212; ' . $this->getTitle();
+            $modalTitle                            = Text::_('LIB_J2COMMERCE_SELECT_USERS') . ' &#8212; ' . $this->getTitle();
             $this->dataAttributes['data-language'] = $language;
         } else {
-            $modalTitle = Text::_('LIB_J2COMMERCE_ARTICLE_SELECT_MODAL_ADD_ARTICLES');
+            $modalTitle = Text::_('LIB_J2COMMERCE_USER_SELECT_MODAL_ADD_USERS');
         }
 
         $this->urls['select']        = (string) $linkItems;
         $this->modalTitles['select'] = $modalTitle;
-        $this->hint                  = $this->hint ?: Text::_('LIB_J2COMMERCE_SELECT_ARTICLES');
+        $this->hint                  = $this->hint ?: Text::_('LIB_J2COMMERCE_SELECT_USERS');
 
-        $this->sql_title_table  = '#__content';
-        $this->sql_title_column = 'title';
+        $this->sql_title_table  = '#__users';
+        $this->sql_title_column = 'name';
         $this->sql_title_key    = 'id';
 
         return $result;
@@ -124,27 +124,27 @@ class ArticleMultiSelectField extends ModalMultiSelectField
         // Get the parent input HTML
         $html = parent::getInput();
 
-        // Add a container for the articles table
+        // Add a container for the users table
         $html .= '<div id="' . $this->id . '_table">';
 
-        // Build the table with the selected articles on first load
+        // Build the table with the selected users on first load
         if (!empty($this->value)) {
-            // Fetch article titles from the database
-            $articles = $this->getValueTitles();
-            $html .= '<div class="my-2"><strong>' . Text::_('LIB_J2COMMERCE_SELECTED_ARTICLES') . ' (' . \count($this->value) . '):</strong></div>';
-            $html .= '<table class="table table-sm table-striped"><thead><tr><th class="w-10">' . Text::_('LIB_J2COMMERCE_ARTICLE_FIELD_ID') . '</th><th>' . Text::_('LIB_J2COMMERCE_ARTICLE_FIELD_NAME') . '</th><th class="text-end w-6"><button type="button" class="btn btn-sm btn-outline-danger" onclick="clearAllItems_' . $this->id . '()" title="' . Text::_('LIB_J2COMMERCE_ARTICLES_CLEAR_ALL') . '"><span class="icon-trash" aria-hidden="true"></span></button></th><th class="w-1"><span class="visually-hidden">' . Text::_('LIB_J2COMMERCE_REMOVE') . '</span></th></tr></thead><tbody>';
-            foreach ($this->value as $index => $articleId) {
-                $articleTitle = isset($articles[$articleId]) ? htmlspecialchars($articles[$articleId]->title, ENT_QUOTES, 'UTF-8') : $articleId;
+            // Fetch user titles from the database
+            $users = $this->getValueTitles();
+            $html .= '<div class="my-2"><strong>' . Text::_('LIB_J2COMMERCE_SELECTED_USERS') . ' (' . \count($this->value) . '):</strong></div>';
+            $html .= '<table class="table table-sm table-striped"><thead><tr><th class="w-10">' . Text::_('LIB_J2COMMERCE_USER_FIELD_ID') . '</th><th>' . Text::_('LIB_J2COMMERCE_USER_FIELD_NAME') . '</th><th class="text-end w-6"><button type="button" class="btn btn-sm btn-outline-danger" onclick="clearAllItems_' . $this->id . '()" title="' . Text::_('LIB_J2COMMERCE_USERS_CLEAR_ALL') . '"><span class="icon-trash" aria-hidden="true"></span></button></th><th class="w-1"><span class="visually-hidden">' . Text::_('LIB_J2COMMERCE_REMOVE') . '</span></th></tr></thead><tbody>';
+            foreach ($this->value as $index => $userId) {
+                $userName = isset($users[$userId]) ? htmlspecialchars($users[$userId]->name, ENT_QUOTES, 'UTF-8') : $userId;
                 $html .= '<tr>';
-                $html .= '<td class="fw-bold">' . $articleId . '</td>';
-                $html .= '<td>' . $articleTitle . '</td>';
-                $html .= '<td class="text-end"><button type="button" class="btn btn-sm btn-danger" onclick="removeItem_' . $this->id . '(' . $articleId . ')" title="' . Text::_('LIB_J2COMMERCE_ARTICLE_CLEAR') . '"><span class="icon-trash" aria-hidden="true"></span></button></td>';
-                $html .= '<td class="w-1"><input type="hidden" name="jform[request][article_ids][' . $index . ']" value="' . $articleId . '" id="' . $this->id . '_hidden_' . $index . '" data-title="' . $articleTitle . '"></td>';
+                $html .= '<td class="fw-bold">' . $userId . '</td>';
+                $html .= '<td>' . $userName . '</td>';
+                $html .= '<td class="text-end"><button type="button" class="btn btn-sm btn-danger" onclick="removeItem_' . $this->id . '(' . $userId . ')" title="' . Text::_('LIB_J2COMMERCE_USER_CLEAR') . '"><span class="icon-trash" aria-hidden="true"></span></button></td>';
+                $html .= '<td class="w-1"><input type="hidden" name="jform[request][user_ids][' . $index . ']" value="' . $userId . '" id="' . $this->id . '_hidden_' . $index . '" data-title="' . $userName . '"></td>';
                 $html .= '</tr>';
             }
             $html .= '</tbody></table>';
         } else {
-            $html .= '<div class="my-2">' . Text::_('LIB_J2COMMERCE_NO_ARTICLES_SELECTED') . '</div>';
+            $html .= '<div class="my-2">' . Text::_('LIB_J2COMMERCE_NO_USERS_SELECTED') . '</div>';
         }
         $html .= '</div>';
 
@@ -171,11 +171,11 @@ class ArticleMultiSelectField extends ModalMultiSelectField
         );
 
         // Load language strings for JavaScript
-        Text::script('LIB_J2COMMERCE_SELECTED_ARTICLES');
-        Text::script('LIB_J2COMMERCE_ARTICLE_FIELD_ID');
-        Text::script('LIB_J2COMMERCE_ARTICLE_FIELD_NAME');
-        Text::script('LIB_J2COMMERCE_ARTICLES_CLEAR_ALL');
-        Text::script('LIB_J2COMMERCE_ARTICLE_CLEAR');
+        Text::script('LIB_J2COMMERCE_SELECTED_USERS');
+        Text::script('LIB_J2COMMERCE_USER_FIELD_ID');
+        Text::script('LIB_J2COMMERCE_USER_FIELD_NAME');
+        Text::script('LIB_J2COMMERCE_USERS_CLEAR_ALL');
+        Text::script('LIB_J2COMMERCE_USER_CLEAR');
         Text::script('LIB_J2COMMERCE_REMOVE');
 
         // Initialize this specific field instance via inline script
@@ -184,7 +184,7 @@ class ArticleMultiSelectField extends ModalMultiSelectField
             if (window.initItemMultiField) {
                 const handler = window.initItemMultiField('{$this->id}');
 
-                // Set up article-specific table update function
+                // Set up user-specific table update function
                 handler.customUpdateTable = function(selectedItems) {
                     const tableContainer = document.getElementById('{$this->id}_table');
                     if (!tableContainer) {
@@ -199,7 +199,7 @@ class ArticleMultiSelectField extends ModalMultiSelectField
                         // Create table structure
                         const caption = handler.createElement('div', {
                             className: 'my-2',
-                            innerHTML: '<strong>' + Joomla.Text._('LIB_J2COMMERCE_SELECTED_ARTICLES') + ' (' + selectedItems.length + '):</strong>'
+                            innerHTML: '<strong>' + Joomla.Text._('LIB_J2COMMERCE_SELECTED_USERS') + ' (' + selectedItems.length + '):</strong>'
                         });
 
                         const table = handler.createElement('table', {
@@ -208,12 +208,12 @@ class ArticleMultiSelectField extends ModalMultiSelectField
 
                         const thead = handler.createElement('thead');
                         const headerRow = handler.createElement('tr', {
-                            innerHTML: '<th class=\"w-10\">' + Joomla.Text._('LIB_J2COMMERCE_ARTICLE_FIELD_ID') + '</th>' +
-                                      '<th>' + Joomla.Text._('LIB_J2COMMERCE_ARTICLE_FIELD_NAME') + '</th>' +
+                            innerHTML: '<th class=\"w-10\">' + Joomla.Text._('LIB_J2COMMERCE_USER_FIELD_ID') + '</th>' +
+                                      '<th>' + Joomla.Text._('LIB_J2COMMERCE_USER_FIELD_NAME') + '</th>' +
                                       '<th class=\"text-end w-6\">' +
                                           '<button type=\"button\" class=\"btn btn-sm btn-outline-danger\" ' +
                                                   'onclick=\"clearAllItems_{$this->id}()\" ' +
-                                                  'title=\"' + Joomla.Text._('LIB_J2COMMERCE_ARTICLES_CLEAR_ALL') + '\">' +
+                                                  'title=\"' + Joomla.Text._('LIB_J2COMMERCE_USERS_CLEAR_ALL') + '\">' +
                                               '<span class=\"icon-trash\" aria-hidden=\"true\"></span>' +
                                           '</button>' +
                                       '</th>' +
@@ -231,7 +231,7 @@ class ArticleMultiSelectField extends ModalMultiSelectField
                             // Create hidden field
                             const hiddenField = handler.createElement('input', {
                                 type: 'hidden',
-                                name: 'jform[request][article_ids][' + index + ']',
+                                name: 'jform[request][user_ids][' + index + ']',
                                 value: item.id,
                                 id: '{$this->id}_hidden_' + index
                             });
@@ -245,7 +245,7 @@ class ArticleMultiSelectField extends ModalMultiSelectField
                                           '<td class=\"text-end\">' +
                                               '<button type=\"button\" class=\"btn btn-sm btn-danger\" ' +
                                                       'onclick=\"removeItem_{$this->id}(' + item.id + ')\" ' +
-                                                      'title=\"' + Joomla.Text._('LIB_J2COMMERCE_ARTICLE_CLEAR') + '\">' +
+                                                      'title=\"' + Joomla.Text._('LIB_J2COMMERCE_USER_CLEAR') + '\">' +
                                                   '<span class=\"icon-trash\" aria-hidden=\"true\"></span>' +
                                               '</button>' +
                                           '</td>';
@@ -281,7 +281,7 @@ class ArticleMultiSelectField extends ModalMultiSelectField
         $data                          = parent::getLayoutData();
         $data['language']              = (string) $this->element['language'];
         $data['multiple']              = true;
-        $data['buttonIcons']['select'] = 'fa-solid fa-file-text';
+        $data['buttonIcons']['select'] = 'fa-solid fa-users';
 
         return $data;
     }
